Refactor MySql client into partial files
