+++ DbaClientX.MySql/MySql.BulkOperations.cs
@@ -1,0 +1,211 @@
+using System;
+using System.Data;
+using System.Threading;
+using System.Threading.Tasks;
+using MySqlConnector;
+
+namespace DBAClientX;
+
+public partial class MySql
+{
+    /// <summary>
+    /// Performs a bulk insert using <see cref="MySqlBulkCopy"/> and the provided <see cref="DataTable"/> payload.
+    /// </summary>
+    public virtual void BulkInsert(
+        string host,
+        string database,
+        string username,
+        string password,
+        DataTable table,
+        string destinationTable,
+        bool useTransaction = false,
+        int? batchSize = null,
+        int? bulkCopyTimeout = null)
+    {
+        if (table == null)
+        {
+            throw new ArgumentNullException(nameof(table));
+        }
+
+        var connectionString = BuildConnectionString(host, database, username, password);
+
+        MySqlConnection? connection = null;
+        var dispose = false;
+        if (useTransaction)
+        {
+            if (_transaction == null || _transactionConnection == null)
+            {
+                throw new DbaTransactionException("Transaction has not been started.");
+            }
+
+            connection = _transactionConnection;
+        }
+        else
+        {
+            connection = CreateConnection(connectionString);
+            OpenConnection(connection);
+            dispose = true;
+        }
+
+        try
+        {
+            var bulkCopy = CreateBulkCopy(connection!, useTransaction ? _transaction : null);
+            bulkCopy.DestinationTableName = destinationTable;
+            if (bulkCopyTimeout.HasValue)
+            {
+                bulkCopy.BulkCopyTimeout = bulkCopyTimeout.Value;
+            }
+
+            foreach (DataColumn column in table.Columns)
+            {
+                bulkCopy.ColumnMappings.Add(new MySqlBulkCopyColumnMapping(column.Ordinal, column.ColumnName, null));
+            }
+
+            if (batchSize.HasValue && batchSize.Value > 0)
+            {
+                var totalRows = table.Rows.Count;
+                for (var offset = 0; offset < totalRows; offset += batchSize.Value)
+                {
+                    var batchTable = table.Clone();
+                    for (var i = offset; i < Math.Min(offset + batchSize.Value, totalRows); i++)
+                    {
+                        batchTable.ImportRow(table.Rows[i]);
+                    }
+
+                    WriteToServer(bulkCopy, batchTable);
+                }
+            }
+            else
+            {
+                WriteToServer(bulkCopy, table);
+            }
+        }
+        catch (Exception ex)
+        {
+            throw new DbaQueryExecutionException("Failed to execute bulk insert.", destinationTable, ex);
+        }
+        finally
+        {
+            if (dispose)
+            {
+                connection?.Dispose();
+            }
+        }
+    }
+
+    /// <summary>
+    /// Performs a bulk insert asynchronously using <see cref="MySqlBulkCopy"/> and the provided <see cref="DataTable"/> payload.
+    /// </summary>
+    public virtual async Task BulkInsertAsync(
+        string host,
+        string database,
+        string username,
+        string password,
+        DataTable table,
+        string destinationTable,
+        bool useTransaction = false,
+        int? batchSize = null,
+        int? bulkCopyTimeout = null,
+        CancellationToken cancellationToken = default)
+    {
+        if (table == null)
+        {
+            throw new ArgumentNullException(nameof(table));
+        }
+
+        var connectionString = BuildConnectionString(host, database, username, password);
+
+        MySqlConnection? connection = null;
+        var dispose = false;
+        if (useTransaction)
+        {
+            if (_transaction == null || _transactionConnection == null)
+            {
+                throw new DbaTransactionException("Transaction has not been started.");
+            }
+
+            connection = _transactionConnection;
+        }
+        else
+        {
+            connection = CreateConnection(connectionString);
+            await OpenConnectionAsync(connection, cancellationToken).ConfigureAwait(false);
+            dispose = true;
+        }
+
+        try
+        {
+            var bulkCopy = CreateBulkCopy(connection!, useTransaction ? _transaction : null);
+            bulkCopy.DestinationTableName = destinationTable;
+            if (bulkCopyTimeout.HasValue)
+            {
+                bulkCopy.BulkCopyTimeout = bulkCopyTimeout.Value;
+            }
+
+            foreach (DataColumn column in table.Columns)
+            {
+                bulkCopy.ColumnMappings.Add(new MySqlBulkCopyColumnMapping(column.Ordinal, column.ColumnName, null));
+            }
+
+            if (batchSize.HasValue && batchSize.Value > 0)
+            {
+                var totalRows = table.Rows.Count;
+                for (var offset = 0; offset < totalRows; offset += batchSize.Value)
+                {
+                    var batchTable = table.Clone();
+                    for (var i = offset; i < Math.Min(offset + batchSize.Value, totalRows); i++)
+                    {
+                        batchTable.ImportRow(table.Rows[i]);
+                    }
+
+                    await WriteToServerAsync(bulkCopy, batchTable, cancellationToken).ConfigureAwait(false);
+                }
+            }
+            else
+            {
+                await WriteToServerAsync(bulkCopy, table, cancellationToken).ConfigureAwait(false);
+            }
+        }
+        catch (Exception ex)
+        {
+            throw new DbaQueryExecutionException("Failed to execute bulk insert.", destinationTable, ex);
+        }
+        finally
+        {
+            if (dispose)
+            {
+                connection?.Dispose();
+            }
+        }
+    }
+
+    /// <summary>
+    /// Creates a configured <see cref="MySqlBulkCopy"/> instance for bulk operations.
+    /// </summary>
+    protected virtual MySqlBulkCopy CreateBulkCopy(MySqlConnection connection, MySqlTransaction? transaction) => new(connection, transaction);
+
+    /// <summary>
+    /// Writes the contents of <paramref name="table"/> to the server using the provided bulk copy instance.
+    /// </summary>
+    protected virtual void WriteToServer(MySqlBulkCopy bulkCopy, DataTable table) => bulkCopy.WriteToServer(table);
+
+    /// <summary>
+    /// Asynchronously writes the contents of <paramref name="table"/> to the server using the provided bulk copy instance.
+    /// </summary>
+    protected virtual Task WriteToServerAsync(MySqlBulkCopy bulkCopy, DataTable table, CancellationToken cancellationToken) => bulkCopy.WriteToServerAsync(table, cancellationToken).AsTask();
+
+    /// <summary>
+    /// Creates a new <see cref="MySqlConnection"/> for the supplied connection string.
+    /// </summary>
+    protected virtual MySqlConnection CreateConnection(string connectionString) => new(connectionString);
+
+    /// <summary>
+    /// Opens a MySQL connection using synchronous APIs.
+    /// </summary>
+    protected virtual void OpenConnection(MySqlConnection connection) => connection.Open();
+
+    /// <summary>
+    /// Opens a MySQL connection asynchronously.
+    /// </summary>
+    protected virtual Task OpenConnectionAsync(MySqlConnection connection, CancellationToken cancellationToken) => connection.OpenAsync(cancellationToken);
+}
