+++ DbaClientX.MySql/MySql.ParallelExecution.cs
@@ -1,0 +1,61 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace DBAClientX;
+
+public partial class MySql
+{
+    /// <summary>
+    /// Executes multiple queries concurrently against the same connection information.
+    /// </summary>
+    public async Task<IReadOnlyList<object?>> RunQueriesInParallel(
+        IEnumerable<string> queries,
+        string host,
+        string database,
+        string username,
+        string password,
+        CancellationToken cancellationToken = default,
+        int? maxDegreeOfParallelism = null)
+    {
+        if (queries == null)
+        {
+            throw new ArgumentNullException(nameof(queries));
+        }
+
+        SemaphoreSlim? throttler = null;
+        if (maxDegreeOfParallelism.HasValue && maxDegreeOfParallelism.Value > 0)
+        {
+            throttler = new SemaphoreSlim(maxDegreeOfParallelism.Value);
+        }
+
+        var tasks = queries.Select(async q =>
+        {
+            if (throttler != null)
+            {
+                await throttler.WaitAsync(cancellationToken).ConfigureAwait(false);
+            }
+
+            try
+            {
+                return await QueryAsync(host, database, username, password, q, cancellationToken: cancellationToken).ConfigureAwait(false);
+            }
+            finally
+            {
+                throttler?.Release();
+            }
+        });
+
+        try
+        {
+            var results = await Task.WhenAll(tasks).ConfigureAwait(false);
+            return results;
+        }
+        finally
+        {
+            throttler?.Dispose();
+        }
+    }
+}
