+++ DbaClientX.MySql/MySql.Transactions.cs
@@ -1,0 +1,177 @@
+using System;
+using System.Data;
+using System.Threading;
+using System.Threading.Tasks;
+using MySqlConnector;
+
+namespace DBAClientX;
+
+public partial class MySql
+{
+    /// <summary>
+    /// Starts a transaction using the default isolation level (<see cref="IsolationLevel.ReadCommitted"/>).
+    /// </summary>
+    public virtual void BeginTransaction(string host, string database, string username, string password)
+        => BeginTransaction(host, database, username, password, IsolationLevel.ReadCommitted);
+
+    /// <summary>
+    /// Starts a transaction with the specified <paramref name="isolationLevel"/>.
+    /// </summary>
+    public virtual void BeginTransaction(string host, string database, string username, string password, IsolationLevel isolationLevel)
+    {
+        lock (_syncRoot)
+        {
+            if (_transaction != null)
+            {
+                throw new DbaTransactionException("Transaction already started.");
+            }
+
+            var connectionString = BuildConnectionString(host, database, username, password);
+
+            _transactionConnection = new MySqlConnection(connectionString);
+            _transactionConnection.Open();
+            _transaction = _transactionConnection.BeginTransaction(isolationLevel);
+        }
+    }
+
+    /// <summary>
+    /// Asynchronously starts a transaction using the default isolation level (<see cref="IsolationLevel.ReadCommitted"/>).
+    /// </summary>
+    public virtual Task BeginTransactionAsync(string host, string database, string username, string password, CancellationToken cancellationToken = default)
+        => BeginTransactionAsync(host, database, username, password, IsolationLevel.ReadCommitted, cancellationToken);
+
+    /// <summary>
+    /// Asynchronously starts a transaction with the specified <paramref name="isolationLevel"/>.
+    /// </summary>
+    public virtual async Task BeginTransactionAsync(string host, string database, string username, string password, IsolationLevel isolationLevel, CancellationToken cancellationToken = default)
+    {
+        lock (_syncRoot)
+        {
+            if (_transaction != null)
+            {
+                throw new DbaTransactionException("Transaction already started.");
+            }
+        }
+
+        var connectionString = BuildConnectionString(host, database, username, password);
+
+        var connection = new MySqlConnection(connectionString);
+        await connection.OpenAsync(cancellationToken).ConfigureAwait(false);
+#if NETSTANDARD2_1_OR_GREATER || NETCOREAPP3_0_OR_GREATER || NET5_0_OR_GREATER
+        var transaction = await connection.BeginTransactionAsync(isolationLevel, cancellationToken).ConfigureAwait(false);
+#else
+        var transaction = connection.BeginTransaction(isolationLevel);
+#endif
+
+        lock (_syncRoot)
+        {
+            if (_transaction != null)
+            {
+                transaction.Dispose();
+                connection.Dispose();
+                throw new DbaTransactionException("Transaction already started.");
+            }
+
+            _transactionConnection = connection;
+            _transaction = transaction;
+        }
+    }
+
+    /// <summary>
+    /// Commits the active transaction.
+    /// </summary>
+    public virtual void Commit()
+    {
+        lock (_syncRoot)
+        {
+            if (_transaction == null)
+            {
+                throw new DbaTransactionException("No active transaction.");
+            }
+
+            _transaction.Commit();
+            DisposeTransactionLocked();
+        }
+    }
+
+    /// <summary>
+    /// Asynchronously commits the active transaction.
+    /// </summary>
+    public virtual async Task CommitAsync(CancellationToken cancellationToken = default)
+    {
+        if (_transaction == null)
+        {
+            throw new DbaTransactionException("No active transaction.");
+        }
+
+        await _transaction.CommitAsync(cancellationToken).ConfigureAwait(false);
+        DisposeTransaction();
+    }
+
+    /// <summary>
+    /// Rolls back the active transaction.
+    /// </summary>
+    public virtual void Rollback()
+    {
+        lock (_syncRoot)
+        {
+            if (_transaction == null)
+            {
+                throw new DbaTransactionException("No active transaction.");
+            }
+
+            _transaction.Rollback();
+            DisposeTransactionLocked();
+        }
+    }
+
+    /// <summary>
+    /// Asynchronously rolls back the active transaction.
+    /// </summary>
+    public virtual async Task RollbackAsync(CancellationToken cancellationToken = default)
+    {
+        if (_transaction == null)
+        {
+            throw new DbaTransactionException("No active transaction.");
+        }
+
+        await _transaction.RollbackAsync(cancellationToken).ConfigureAwait(false);
+        DisposeTransaction();
+    }
+
+    private void DisposeTransaction()
+    {
+        lock (_syncRoot)
+        {
+            DisposeTransactionLocked();
+        }
+    }
+
+    private void DisposeTransactionLocked()
+    {
+        _transaction?.Dispose();
+        _transaction = null;
+        _transactionConnection?.Dispose();
+        _transactionConnection = null;
+    }
+
+    /// <inheritdoc />
+    protected override bool IsTransient(Exception ex) =>
+        ex is MySqlException mysqlEx &&
+        mysqlEx.ErrorCode is MySqlErrorCode.ConnectionCountError
+            or MySqlErrorCode.LockDeadlock
+            or MySqlErrorCode.LockWaitTimeout
+            or MySqlErrorCode.UnableToConnectToHost
+            or MySqlErrorCode.XARBDeadlock;
+
+    /// <inheritdoc />
+    protected override void Dispose(bool disposing)
+    {
+        if (disposing)
+        {
+            DisposeTransaction();
+        }
+
+        base.Dispose(disposing);
+    }
+}
